// src/utility/enums/entity.enum.ts
export enum Entity {
  SOFTWARE_APPLICATION = 'SoftwareApplication',
  SOFTWARE_SOURCE_CODE = 'SoftwareSourceCode',
  CONTACT_POINT = 'ContactPoint',
  DATA_PRODUCT = 'DataProduct',
  DISTRIBUTION = 'Distribution',
  DOCUMENTATION = 'Documentation',
  GROUP = 'Group',
  IDENTIFIER = 'Identifier',
  LOCATION = 'Location',
  MAPPING = 'Mapping',
  OPERATION = 'Operation',
  ORGANIZATION = 'Organization',
  PERIOD_OF_TIME = 'PeriodOfTime',
  PERSON = 'Person',
  USER = 'User',
  WEBSERVICE = 'WebService',
  CATEGORY = 'Category',
}